fix: 동적 라우터 수정
