+++ src/router/lazy.ts
@@ -5,7 +5,9 @@ export const BlindMatchPage = lazy(
 );
 export const BoothPage = lazy(() => import('@pages/booth/booth'));
 export const HomePage = lazy(() => import('@pages/home/home'));
-export const HomeDetailPage = lazy(() => import('@pages/home/detail/[id]'));
+export const HomeDetailPage = lazy(
+  () => import('@pages/home-detail/home-detail'),
+);
 export const LandPage = lazy(() => import('@pages/land/land'));
 export const LoginPage = lazy(() => import('@pages/login/login'));
 export const LostItemsPage = lazy(() => import('@pages/lost-items/lost-items'));
